Refactor the iteration of tokens from the lexer
Previously I used a buffer (vector) where I read all the tokens from the string and used the iterator from the vector to go over the tokens. This had the disadvantage that when you got to the end of the iterator, you'd get C++ assertions which aren't really catchable. Therefore I had a lot of duplicated error handling code.

My new approach uses a streaming fashion to read the tokens from the program input and abstracts away a lot of exception throwing and error handling while at the same time giving more consistent error messages. Of course you still have some assertion and exception throwing in the parsing code, but it is much more clean now. It saves almost 200 lines of code.