Separate the implementation of type details and conversions to type implementations
This allows us not to worry about the type conversion details in the evaluation visitors but let it happen through the implementation of the specific type instead. It also completely decouples us from the built-in AST type and allows us to use the symbol lookup table for each variable and result type.