Support string concatenation by refactoring the type system
I was not happy with the type system at all because it was dicy and hard to adapt. There is still a memory leak to fix in the string concatenation because it allocates a string but never registers it with the runtime, so it may become a dangling pointer later.