Fix recursive calls in functions
Apparently arguments where evaluated in the unconstructed scope of the procedure that is being prepared. This of course caused the arguments to be evaluated with uninitialized memory (actually the memory table is initialized with zeroed bytes) which caused invalid evaluation of the program.

I added sample programs for Fibonacci. The interpreter can now evaluate the 20th number of fibonacci in 4 seconds.